+++ src/Zicht/Bundle/UrlBundle/DependencyInjection/ZichtUrlExtension.php
@@ -5,12 +5,13 @@
 
 namespace Zicht\Bundle\UrlBundle\DependencyInjection;
 
-use Symfony\Component\DependencyInjection\Alias;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\Alias;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Zicht\Bundle\UrlBundle\Url\ShortUrlManager;
 
 /**
  * DI Extension for the URL services.
@@ -20,7 +21,7 @@ class ZichtUrlExtension extends Extension
     /**
      * Responds to the twig configuration parameter.
      *
-     * @param array            $configs
+     * @param array $configs
      * @param ContainerBuilder $container
      * @return void
      */
@@ -63,6 +64,10 @@ class ZichtUrlExtension extends Extension
 
         if ($container->hasDefinition('zicht_url.aliasing')) {
             $container->getDefinition('zicht_url.twig_extension')->addArgument(new Reference('zicht_url.aliasing'));
+        }
+
+        if ($container->hasDefinition(ShortUrlManager::class)) {
+            $container->getDefinition(ShortUrlManager::class)->addArgument(new Reference('zicht_url.aliasing'));
         }
 
         if ($container->hasDefinition('zicht_url.mapper.html')) {